requirements.txt
requests
pydantic
aiohttp
python-dotenv 
semantic-kernel[azure] 